+++ CHANGELOG.md
@@ -17,9 +17,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add "Add Next" and "Add Previous" analysis in analysis node. [#6](https://github.com/IN-CORE/incore-studio/issues/6)
 - Move dependency_graph.json to public and fetch it so it can be overwritten by helm. [#33](https://github.com/IN-CORE/incore-studio/issues/33)
 - Add analysis configuration panel. [#31](https://github.com/IN-CORE/incore-studio/issues/31)
+- Workflow Execution Summary page. [#16](https://github.com/IN-CORE/incore-studio/issues/16)
 - Error Handling for missing dependency graph or Valid tools in workflow. [#50](https://github.com/IN-CORE/incore-studio/issues/50)
 - Create visualization components [#18](https://github.com/IN-CORE/incore-studio/issues/18)
 - Create empty project [#42](https://github.com/IN-CORE/incore-studio/issues/42)
+- Modal that adds resources from service to project [#40](https://github.com/IN-CORE/incore-studio/issues/40)
 
 ### Changed